cutting_edge_salon/requirements.txt
customtkinter
packaging
pillow
tkcalendar